+++ cloudant-client/src/test/java/com/cloudant/tests/HttpTest.java
@@ -9,6 +9,7 @@ import static org.junit.Assert.fail;
 
 import com.cloudant.client.api.CloudantClient;
 import com.cloudant.client.org.lightcouch.CouchDbException;
+import com.cloudant.client.org.lightcouch.TooManyRequestsException;
 import com.cloudant.http.Http;
 import com.cloudant.http.HttpConnection;
 import com.cloudant.http.HttpConnectionInterceptorContext;
@@ -16,6 +17,7 @@ import com.cloudant.http.HttpConnectionRequestInterceptor;
 import com.cloudant.http.HttpConnectionResponseInterceptor;
 import com.cloudant.http.interceptors.BasicAuthInterceptor;
 import com.cloudant.http.interceptors.CookieInterceptor;
+import com.cloudant.http.interceptors.RequestLimitInterceptor;
 import com.cloudant.test.main.RequiresCloudant;
 import com.cloudant.tests.util.CloudantClientResource;
 import com.cloudant.tests.util.DatabaseResource;
@@ -23,10 +25,12 @@ import com.cloudant.tests.util.MockWebServerResources;
 import com.cloudant.tests.util.Utils;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
+import com.squareup.okhttp.mockwebserver.Dispatcher;
 import com.squareup.okhttp.mockwebserver.MockResponse;
 import com.squareup.okhttp.mockwebserver.MockWebServer;
 import com.squareup.okhttp.mockwebserver.RecordedRequest;
 
+import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
@@ -38,6 +42,7 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.lang.reflect.Field;
 import java.net.URL;
 import java.net.URLDecoder;
 import java.nio.charset.Charset;
@@ -562,5 +567,126 @@ public class HttpTest {
         public InputStream getInputStream() throws IOException {
             return new ByteArrayInputStream(content);
         }
+    }
+
+    private MockResponse response429 = new MockResponse().setResponseCode(429).setBody("{\"error" +
+            "\":\"too_many_requests\", \"reason\":\"example reason\"}\r\n");
+
+    /**
+     * For testing purposes reduce the default backoff time in the RequestLimitInterceptor
+     */
+    @BeforeClass
+    public static void reduceBackoffTime() throws Exception {
+        // Reflectively reduce the time limit for testing
+        Field f = RequestLimitInterceptor.class.getDeclaredField("initialSleep");
+        f.setAccessible(true);
+        f.set(null, 1); // Set the value to 1 ms
+    }
+
+    /**
+     * Test that a request is replayed in response to a 429.
+     *
+     * @throws Exception
+     */
+    @Test
+    public void test429Backoff() throws Exception {
+        mockWebServer.enqueue(response429);
+        mockWebServer.enqueue(new MockResponse());
+
+        CloudantClient c = CloudantClientHelper.newMockWebServerClientBuilder(mockWebServer)
+                .build();
+        c.executeRequest(Http.GET(c.getBaseUri()));
+
+        assertEquals("There should be 2 requests", 2, mockWebServer.getRequestCount());
+    }
+
+    /**
+     * Test that the maximum number of retries is reached and the backoff is of at least the
+     * expected duration.
+     *
+     * @throws Exception
+     */
+    @Test
+    public void test429BackoffMaxDefault() throws Exception {
+
+        // Always respond 429 for this test
+        mockWebServer.setDispatcher(new Dispatcher() {
+            @Override
+            public MockResponse dispatch(RecordedRequest request) throws InterruptedException {
+                return response429;
+            }
+        });
+
+
+        long startTime = System.currentTimeMillis();
+        try {
+            CloudantClient c = CloudantClientHelper.newMockWebServerClientBuilder(mockWebServer)
+                    .build();
+            c.executeRequest(Http.GET(c.getBaseUri()));
+        } catch (TooManyRequestsException e) {
+            long duration = System.currentTimeMillis() - startTime;
+            // 9 backoff periods for 10 attempts: 1 + 2 + 4 + 8 + 16 + 32 + 64 + 128 + 256 = 511 ms
+            assertTrue("The duration should be at least 511 ms, but was " + duration, duration >=
+                    511);
+            assertEquals("There should be 10 request attempts", 10, mockWebServer
+                    .getRequestCount());
+        }
+
+    }
+
+    /**
+     * Test that the number of retries is configurable.
+     *
+     * @throws Exception
+     */
+    @Test
+    public void test429BackoffMaxConfigured() throws Exception {
+
+        // Always respond 429 for this test
+        mockWebServer.setDispatcher(new Dispatcher() {
+            @Override
+            public MockResponse dispatch(RecordedRequest request) throws InterruptedException {
+                return response429;
+            }
+        });
+
+
+        long startTime = System.currentTimeMillis();
+        try {
+            CloudantClient c = CloudantClientHelper.newMockWebServerClientBuilder(mockWebServer)
+                    .build();
+            c.executeRequest(Http.GET(c.getBaseUri()).setNumberOfRetries(3));
+        } catch (TooManyRequestsException e) {
+            long duration = System.currentTimeMillis() - startTime;
+            // 2 backoff periods for 3 attempts: 1 + 2 = 3 ms
+            assertTrue("The duration should be at least 3 ms, but was " + duration, duration >=
+                    3);
+            assertEquals("There should be 3 request attempts", 3, mockWebServer
+                    .getRequestCount());
+        }
+
+    }
+
+    /**
+     * Test that an integer number of seconds delay specified by a Retry-After header is honoured.
+     *
+     * @throws Exception
+     */
+    @Test
+    public void test429BackoffRetryAfter() throws Exception {
+
+        mockWebServer.enqueue(response429.addHeader("Retry-After", "1"));
+        mockWebServer.enqueue(new MockResponse());
+
+        long startTime = System.currentTimeMillis();
+        CloudantClient c = CloudantClientHelper.newMockWebServerClientBuilder(mockWebServer)
+                .build();
+        c.executeRequest(Http.GET(c.getBaseUri()));
+
+        long duration = System.currentTimeMillis() - startTime;
+        assertTrue("The duration should be at least 1000 ms, but was " + duration, duration >=
+                1000);
+        assertEquals("There should be 2 request attempts", 2, mockWebServer
+                .getRequestCount());
     }
 }